% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMultioviz.R
\name{runMultioviz}
\alias{runMultioviz}
\title{Run Shiny App}
\usage{
runMultioviz(X = NULL, y = NULL, mask = NULL, userScript = NULL)
}
\arguments{
\item{X}{N x J matrix where N is the number of samples and J is the size of the set of molecular variables for molecular level one}

\item{y}{N-dimensional matrix of quantitative traits}

\item{mask}{J x G matrix of pre-defined annotations where J is the number of molecular variables for molecular level 1 and G is the number of molecular variables for molecular level 2}
}
\description{
This function takes in parameters for a mathematical model and the wrapper script for the model and runs the multioviz Shiny application
}
